+++ main-server/src/db/telemetry/index.js
@@ -118,6 +118,21 @@ export async function addTendermintAndABCIVersionLog(
   }
 }
 
+export async function addProcessLog(nodeId, { node_id, log }) {
+  try {
+    await getRedis().xadd(
+      `${nodeId}:process-logs`,
+      '*',
+      'node_id',
+      node_id,
+      'log',
+      log
+    );
+  } catch (err) {
+    logger.error({ err });
+  }
+}
+
 export async function addNewRequestEvent(
   nodeId,
   { request_id, node_id, state_code, source_timestamp, additional_data }